Refactor file paths for canvas setup and circuit components
